Refactorings and locking for all public filesystem actions to avoid concurrency issues

+++ Source/CarthageKit/FileLock.swift
@@ -10,6 +10,7 @@ import Result
 import ReactiveSwift
 import ReactiveTask
 
+/// Protocol describing the public interface for concrete implementations of lock
 protocol Lock {
     
     /// Whether or not the lock is currently locked
@@ -43,10 +44,14 @@ extension Lock {
 final class FileLock: Lock {
     
     private static let retryInterval = 1.0
+    private var wasLocked = false
     let lockFileURL: URL
+    let isRecursive: Bool
+    var onWait: ((FileLock) -> Void)?
     
-    init(lockFileURL: URL) {
+    init(lockFileURL: URL, isRecursive: Bool = true) {
         self.lockFileURL = lockFileURL
+        self.isRecursive = isRecursive
     }
     
     deinit {
@@ -55,20 +60,30 @@ final class FileLock: Lock {
     
     /// Tries a lock with an optional timeoutDate. If the lock was acquired before the timeout date true will be returned, false otherwise.
     func lock(timeoutDate: Date?) -> Bool {
+        var waiting = false
         while true {
+            let processId = self.processId
+            if self.isRecursive && processId == self.lockingProcessId {
+                return true
+            }
             guard let _ = try? FileManager.default.createDirectory(at: lockFileURL.deletingLastPathComponent(), withIntermediateDirectories: true) else {
                 return false
             }
             let task = Process()
             task.launchPath = "/usr/bin/shlock"
-            task.arguments = ["-f", lockFileURL.path, "-p", processId]
+            task.arguments = ["-f", lockFileURL.path, "-p", String(processId)]
             task.launch()
             task.waitUntilExit()
             if task.terminationStatus == 0 {
+                wasLocked = true
                 return true
             }
             if timeoutDate.map({ $0.timeIntervalSinceNow <= 0 }) ?? true {
                 break
+            }
+            if !waiting {
+                onWait?(self)
+                waiting = true
             }
             Thread.sleep(forTimeInterval: FileLock.retryInterval)
         }
@@ -87,10 +102,23 @@ final class FileLock: Lock {
         task.waitUntilExit()
         return task.terminationStatus == 0
     }
+
+    var lockingProcessId: Int? {
+        guard isLocked else {
+            return nil
+        }
+        guard let contents = try? String(contentsOfFile: lockFileURL.path) else {
+            return nil
+        }
+        return Int(contents.trimmingCharacters(in: .whitespacesAndNewlines))
+    }
     
     /// Unlocks the lock, returns true if lock was released, false otherwise (e.g. because the lock file did not exist anymore).
     @discardableResult
     func unlock() -> Bool {
+        guard wasLocked else {
+            return false
+        }
         do {
             try FileManager.default.removeItem(at: self.lockFileURL)
             return true
@@ -99,8 +127,8 @@ final class FileLock: Lock {
         }
     }
     
-    private var processId: String {
-        return String(ProcessInfo.processInfo.processIdentifier)
+    private var processId: Int {
+        return Int(ProcessInfo.processInfo.processIdentifier)
     }
 }
 
@@ -116,14 +144,26 @@ final class URLLock: Lock {
     
     let url: URL
     private let fileLock: FileLock
+    var onWait: ((URLLock) -> Void)? {
+        didSet {
+            fileLock.onWait = { [weak self] fileLock in
+                guard let self = self else { return }
+                self.onWait?(self)
+            }
+        }
+    }
     
-    convenience init(url: URL, lockFileNamingStrategy: (URL) -> URL = URLLock.defaultLockFileNamingStrategy) {
-        self.init(url: url, lockFileURL: lockFileNamingStrategy(url))
+    convenience init(url: URL, isRecursive: Bool = true, lockFileNamingStrategy: (URL) -> URL = URLLock.defaultLockFileNamingStrategy) {
+        self.init(url: url, lockFileURL: lockFileNamingStrategy(url), isRecursive: isRecursive)
     }
 
-    init(url: URL, lockFileURL: URL) {
+    convenience init(url: URL, lockFileURL: URL, isRecursive: Bool = true) {
+        self.init(url: url, fileLock: FileLock(lockFileURL: lockFileURL, isRecursive: isRecursive))
+    }
+
+    init(url: URL, fileLock: FileLock) {
         self.url = url
-        self.fileLock = FileLock(lockFileURL: lockFileURL)
+        self.fileLock = fileLock
     }
     
     func lock(timeoutDate: Date?) -> Bool {
@@ -137,5 +177,20 @@ final class URLLock: Lock {
     @discardableResult
     func unlock() -> Bool {
         return fileLock.unlock()
+    }
+}
+
+extension URLLock {
+    static var globalWaitHandler: ((URLLock) -> Void)?
+
+    static func lockReactive(url: URL, timeout: Int? = nil, onWait: ((URLLock) -> Void)? = URLLock.globalWaitHandler) -> SignalProducer<URLLock, CarthageError> {
+        return SignalProducer({ () -> Result<URLLock, CarthageError> in
+            let lock = URLLock(url: url)
+            lock.onWait = onWait
+            guard lock.lock(timeout: timeout == nil ? TimeInterval(Int.max) : TimeInterval(timeout!)) else {
+                return .failure(CarthageError.lockError(url: url, timeout: timeout))
+            }
+            return .success(lock)
+        })
     }
 }
